backend/requirements.txt
Django==4.2.4
gunicorn==21.2.0
djangorestframework==3.14.0